meal-tracker/backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Pillow
python-dotenv==1.0.1
supabase==2.6.0
